Adjust code review issues
